task manager main + waiter
* Have a task manager lock + a waiter lock. This makes is so that there
is a process executing the task manager schedule() + another dispatcher
process waiting to go next. The waiting lock is used as a signal. New
jobs or new events that happen that MIGHT trigger the task manager can
ask "Is there already a task manager process that will run for me?". If
so, then don't schedule another one. If not, then schedule, that will be
the waiter.
* The above is the ideal. What actually ends up happening. Given the
scenario where you start 20 jobs in a loop. The 19ish jobs see that
there is no waiter and submit to be the waiter. This is counter
productive as those 19 jobs will become waiters and we only actually
want 1 waiter at a time.
* Fix that I'd have to think about more: Can we do a trylock on the
waiter when we are in the dispatcher task for schedule and bail out?
* Totally different idea: Store in the database the fact that there is a
task manager & a waiter. When deciding to submit a new task manager run.
If there is a waiter, don't submit.